+++ web/controller/api.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/mhsanaei/3x-ui/v3/web/middleware"
 	"github.com/mhsanaei/3x-ui/v3/web/service"
+	"github.com/mhsanaei/3x-ui/v3/web/service/panel"
 	"github.com/mhsanaei/3x-ui/v3/web/session"
 
 	"github.com/gin-gonic/gin"
@@ -20,8 +21,8 @@ type APIController struct {
 	settingController     *SettingController
 	xraySettingController *XraySettingController
 	settingService        service.SettingService
-	userService           service.UserService
-	apiTokenService       service.ApiTokenService
+	userService           panel.UserService
+	apiTokenService       panel.ApiTokenService
 	Tgbot                 service.Tgbot
 }
