+++ web/controller/server.go
@@ -14,6 +14,7 @@ import (
 	"github.com/mhsanaei/3x-ui/v3/web/entity"
 	"github.com/mhsanaei/3x-ui/v3/web/global"
 	"github.com/mhsanaei/3x-ui/v3/web/service"
+	"github.com/mhsanaei/3x-ui/v3/web/service/panel"
 	"github.com/mhsanaei/3x-ui/v3/web/websocket"
 
 	"github.com/gin-gonic/gin"
@@ -27,7 +28,7 @@ type ServerController struct {
 
 	serverService      service.ServerService
 	settingService     service.SettingService
-	panelService       service.PanelService
+	panelService       panel.PanelService
 	xrayMetricsService service.XrayMetricsService
 }
