+++ web/service/panel/api_token.go
@@ -1,4 +1,4 @@
-package service
+package panel
 
 import (
 	"crypto/subtle"
