+++ web/service/panel/panel_test.go
@@ -1,4 +1,4 @@
-package service
+package panel
 
 import "testing"
